+++ library/src/test/kotlin/it/krzeminski/githubactions/actions/CustomActionTest.kt
@@ -2,8 +2,6 @@ package it.krzeminski.githubactions.actions
 
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
-import it.krzeminski.githubactions.domain.ExternalActionStep
-import it.krzeminski.githubactions.yaml.stepsToYaml
 
 class CustomActionTest : FunSpec({
 
@@ -18,24 +16,6 @@ class CustomActionTest : FunSpec({
                 "compiler" to "latexmk",
             )
         )
-        val steps = listOf(
-            ExternalActionStep(
-                id = "latex",
-                name = "Latex",
-                action = customAction
-            )
-        )
-
-        // when
-        val yaml = steps.stepsToYaml()
-
-        // then
-        yaml shouldBe """|- id: latex
-                         |  name: Latex
-                         |  uses: xu-cheng/latex-action@v2
-                         |  with:
-                         |    root_file: report.tex
-                         |    compiler: latexmk""".trimMargin()
 
         // given
         val outputs = customAction.buildOutputObject("someStepId")
